Add compression tests
Adds a test_compression test which tests applying operations on
compressed data. Compression is applied only to the numeric data, with
any data before the offset or after the numeric data uncompressed random
bytes.

The list of compression algorithms to test is defined by
COMPRESSION_ALGS in config.py, which by default includes gzip and zlib.
For testing older versions of the active storage server,
COMPRESSION_ALGS may be set to an empty list.